% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessment_longitudinal_table.R
\name{assessment_longitudinal_table}
\alias{assessment_longitudinal_table}
\title{Longitudinal Table of Standardized Scores}
\usage{
assessment_longitudinal_table(
  dat,
  id = "NACCID",
  date = "VISITDATE",
  descriptions = c(Impaired = 0.03, Borderline = 0.1, `Low Average` = 0.26, Average =
    0.76, `High Average` = 0.92, Superior = 0.97, `Very Superior` = 1),
  fill_values,
  methods = "infer",
  table_font_size = 100,
  table_id = NULL,
  show_all_visits = T,
  stubhead_label = NULL
)
}
\arguments{
\item{dat}{Data to use. Must have exactly one row (data from one participant and one visit).}

\item{id}{Column containing ID.}

\item{date}{column with visit dates.}

\item{descriptions}{A named vector giving cutoffs to use for creating descriptions.}

\item{fill_values}{(optional) A named vector of same length (with same names)
as \code{descriptions} with hex color values to use. If \code{NULL}, no colorcoding used.
By default, evenly spread out colors from red through yellow to green are used.}

\item{methods}{(optional) either list of named entries specifying which model was used
for standardizing cognitive scores, or the character string "infer". For the
latter, methods are infered from the \code{dat} object using
\code{NpsychBatteryNorms::methods_from_std_data}. If specified, footnotes are added
to indicate the methods used.}

\item{table_font_size}{(optional) scalar indicating font size as a percentage.}

\item{table_id}{character string to add as id to table.}

\item{show_all_visits}{logical; should all visits be displayed (\code{TRUE}, default), or
only visits with at least one standardized score (\code{FALSE})?}

\item{stubhead_label}{passed to the label argument of \code{gt::tab_stubhead}.}
}
\description{
Create a table of standardized scores across multiple visits.
}
