% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarkerModule.R
\name{biomarkerUI}
\alias{biomarkerUI}
\alias{biomarkerServer}
\alias{biomarkerApp}
\title{Biomarker UI}
\usage{
biomarkerUI(id)

biomarkerServer(
  id,
  adrc_ptid,
  biomarker_api,
  use_mirai = rlang::is_installed("mirai")
)

biomarkerApp(adrc_ptid, biomarker_api)
}
\arguments{
\item{id}{A string used to namespace the module.}

\item{adrc_ptid}{ADRC ptid. A character vector with patient ID's that can be chosen.}

\item{biomarker_api}{Panda API key.}

\item{use_mirai}{A logical value specifying if \code{"mirai"} should be . Optional; defaults to \code{TRUE} if \code{"mirai"} is installed.}
}
\value{
A UI definition.

NULL.

A \code{shinyApp} object.
}
\description{
A short description...

A short description...

A short description...
}
\keyword{internal}
