% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birth_vars.R
\docType{data}
\name{birth_vars}
\alias{birth_vars}
\title{Variables necessary for Birth Dates}
\format{
Character vector of variables used for determining birth date.
}
\usage{
birth_vars
}
\description{
Variables necessary for Birth Dates
}
\keyword{datasets}
