% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_fill_colors.R
\name{calc_fill_colors}
\alias{calc_fill_colors}
\title{Calculate Fill Colors from Number of Descriptions}
\usage{
calc_fill_colors(n)
}
\arguments{
\item{n}{Number of colors to return}
}
\description{
This is almost equivalent to \code{RColorBrewer::brewer.pal(n, "RdYlGn")}
}
