% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_line_colors.R
\name{calc_line_colors}
\alias{calc_line_colors}
\title{Colors for Lines on Plotly Charts}
\usage{
calc_line_colors(n)
}
\arguments{
\item{n}{number of lines we need colors for}
}
\description{
Equivalent to \code{RColorBrewer::brewer.pal(n, "Paired")}
}
