% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_colors.R
\name{check_colors}
\alias{check_colors}
\title{Check colors}
\usage{
check_colors(x, return_non_colors = F)
}
\arguments{
\item{x}{A vector or list of elements}

\item{return_non_colors}{Optional. If \code{TRUE}, return all entries that are not recognized as hex colors as a named character vector where names give the entry number. If \code{FALSE} (default), return logical indicating if all entries are recognized as hex colors or not.}
}
\value{
Either a logical (if \code{return_non_colors} is \code{FALSE}) where \code{TRUE} indicates that all elements of \code{x} are valid hex colors, or a named character vector.
}
\description{
Check if all elements of a vector or list are valid hex colors.
}
