% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colSelectModule.R
\name{colSelectUI}
\alias{colSelectUI}
\alias{colSelectServer}
\alias{colSelectApp}
\title{Shiny Module for Selecting Columns}
\usage{
colSelectUI(id)

colSelectServer(
  id,
  col_names,
  default_methods,
  data_type,
  allow_col_selection = T
)

colSelectApp(
  col_names,
  default_methods,
  data_type = c("nacc", "wls", "wadrc"),
  allow_col_selection = T
)
}
\arguments{
\item{id}{id to link shiny modules}

\item{col_names}{Column names.}

\item{default_methods}{Default methods.}

\item{data_type}{One of \code{"nacc"}, \code{"wls"}, or \code{"wadrc"}.}

\item{allow_col_selection}{A logical; whether column selection is allowed. If
\code{TRUE},}
}
\value{
A shiny app.
}
\description{
This shiny module gives the user the option to manually map column names to
NACC variable names.

A short description...
}
