% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/critical_vars.R
\docType{data}
\name{critical_vars}
\alias{critical_vars}
\title{Variables Necessary}
\format{
Character vector of variables necessary for standardization.
}
\usage{
critical_vars
}
\description{
Variables Necessary
}
\keyword{datasets}
