% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswalk_translations.R
\docType{data}
\name{crosswalk_translations}
\alias{crosswalk_translations}
\title{Crosswalk Variables}
\format{
Named character vector. Each entry and its name are plotted in the same line,
and on the same line in the longitudinal table.
}
\usage{
crosswalk_translations
}
\description{
Crosswalk Variables
}
\keyword{datasets}
