% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSelectModule.R
\name{dataSelectUI}
\alias{dataSelectUI}
\alias{dataSelectServer}
\alias{dataSelectApp}
\title{Shiny module to select data for assessment table}
\usage{
dataSelectUI(id)

dataSelectServer(id)

dataSelectApp()
}
\arguments{
\item{id}{id to link shiny modules}
}
\description{
Shiny module to select data for assessment table
}
