% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_data.R
\docType{data}
\name{demo_data}
\alias{demo_data}
\title{Demo Data}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 352 rows and 158 columns.
}
\usage{
demo_data
}
\description{
A completely made up demo data set meant to mimic the structure of
the NACC data. While IDs look like valid NACC IDs, these are simply the
characters "NACC" with six random digits. Similarly, birth and visit dates
are randomly created, and so are all scores with the only constraint that
-4's are believable. (For example, if OTRAILA is -4, so is OTRAILB.)
}
\keyword{datasets}
