% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptionsModule.R
\name{descriptionsServer}
\alias{descriptionsServer}
\alias{descriptionsApp}
\title{Descriptions Server}
\usage{
descriptionsServer(
  id,
  default_descriptions = c(Impaired = 0.03, Borderline = 0.1, `Low Average` = 0.26,
    Average = 0.76, `High Average` = 0.92, Superior = 0.97, `Very Superior` = 1)
)

descriptionsApp(
  default_descriptions = c(Impaired = 0.03, Borderline = 0.1, `Low Average` = 0.26,
    Average = 0.76, `High Average` = 0.92, Superior = 0.97, `Very Superior` = 1)
)
}
\arguments{
\item{id}{A string.}

\item{default_descriptions}{Optional. A named numeric vector with names for \code{Label} and values for \code{Upper_Bound}.}
}
\value{
A list containing \code{fill_values} and \code{descriptions}.
}
\description{
A short description...
}
