% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptionsModule.R
\name{descriptionsUI}
\alias{descriptionsUI}
\title{Descriptions UI}
\usage{
descriptionsUI(id)
}
\arguments{
\item{id}{A string used to namespace the module.}
}
\value{
A UI definition.
}
\description{
UI to let the user select the groups used for coloring. Includes a table where
group labels can be specified, cutoffs set, and color selected.
}
