% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag_contr_pairs.R
\docType{data}
\name{diag_contr_pairs}
\alias{diag_contr_pairs}
\title{Etiology Specific Variables}
\format{
Tibble with four columns: \code{presump_etio_diag} contains variable names that indicate if disease is present; \code{contribution} contains variable names that indicate if disease contributes to diagnosis; \code{other} gives variable names that give description of other diagnoses; \code{disease} gives human readable names for diseases.
}
\usage{
diag_contr_pairs
}
\description{
Etiology Specific Variables
}
\keyword{datasets}
