% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_measures.R
\name{functional_measures_table}
\alias{functional_measures_table}
\title{Summary Table with Functional Measures}
\usage{
functional_measures_table(dat, age, orientation = "h")
}
\arguments{
\item{dat}{Data set to use; see \code{demo_data} for example}

\item{age}{string specifying column containing age}

\item{orientation}{String specifying if dates should be displayed
horizontally as columns ("h") or vertically as rows ("v")}
}
\description{
Summary Table with Functional Measures
}
\examples{
prep_dat <- prepare_data(demo_data)
functional_measures_table(
  dat = prep_dat[prep_dat$NACCID == prep_dat$NACCID[1],]
)

}
