% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longTableModule.R
\name{longTableUI}
\alias{longTableUI}
\alias{longTableServer}
\title{Shiny Module to Display Longitudinal Trends in Table}
\usage{
longTableUI(id)

longTableServer(
  id,
  dat,
  methods = "infer",
  table_font_size = shiny::reactive(100),
  print_updating = F
)
}
\arguments{
\item{id}{string to tie UI to Server}

\item{dat}{Data to use. Must have exactly one row (data from one participant and one visit).}

\item{methods}{(optional) either list of named entries specifying which model was used
for standardizing cognitive scores, or the character string "infer". For the
latter, methods are infered from the \code{dat} object using
\code{NpsychBatteryNorms::methods_from_std_data}. If specified, footnotes are added
to indicate the methods used.}

\item{table_font_size}{(optional) scalar indicating font size as a percentage.}

\item{print_updating}{logical (default \code{FALSE}); should message be printed to let user know the table is updating. For debugging.}
}
\description{
Table version of the plotCogVarModule.
}
