% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nacc_to_wadrc.R
\docType{data}
\name{nacc_to_wadrc}
\alias{nacc_to_wadrc}
\title{Translate NACC variable names to WADRC variable names}
\format{
A named vector where names of entries correspond to NACC variable
names, and entries contain WADRC variable names.
}
\usage{
nacc_to_wadrc
}
\description{
Translate NACC variable names to WADRC variable names
}
\keyword{datasets}
