% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nacc_to_wls.R
\docType{data}
\name{nacc_to_wls}
\alias{nacc_to_wls}
\title{Translate NACC variable names to WLS variable names}
\format{
A named vector where names of entries correspond to NACC variable
names, and entries contain WLS variable names.
}
\usage{
nacc_to_wls
}
\description{
Translate NACC variable names to WLS variable names
}
\keyword{datasets}
