% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nacc_var_groups.R
\docType{data}
\name{nacc_var_groups}
\alias{nacc_var_groups}
\title{NACC Variable Groupings}
\format{
A named character vector where names are NACC variables, and entries are
corresponding groups.
}
\usage{
nacc_var_groups
}
\description{
A named character vector giving groupings of NACC variables used in
\code{\link[=assessment_summary_table]{assessment_summary_table()}}
}
\keyword{datasets}
