% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nacc_var_labels.R
\docType{data}
\name{nacc_var_labels}
\alias{nacc_var_labels}
\title{Variable Labels for NACC Variables}
\format{
Named character vector with names corresponding to NACC variable names
and entries with pretty labels.
}
\usage{
nacc_var_labels
}
\description{
Variable Labels for NACC Variables
}
\keyword{datasets}
