% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NpsychAssessmentTool-package.R
\docType{package}
\name{NpsychAssessmentTool-package}
\alias{NpsychAssessmentTool}
\alias{NpsychAssessmentTool-package}
\title{NpsychAssessmentTool: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rmtrane/NpsychAssessmentTool}
  \item Report bugs at \url{https://github.com/rmtrane/NpsychAssessmentTool/issues}
}

}
\author{
\strong{Maintainer}: Ralph Møller Trane \email{rtrane@wisc.edu}

}
\keyword{internal}
