% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onRender.R
\name{onRender}
\alias{onRender}
\title{See \code{help(htmlwidgets::onRender)}}
\usage{
onRender(x, jsCode, data = NULL)
}
\arguments{
\item{x}{HTML Widget object}

\item{jsCode}{Character vector containing JavaScript code}

\item{data}{Additional argument passed to jsCode. Can be any R object that can be serialized to JSON.}
}
\description{
See \code{help(htmlwidgets::onRender)}
}
