% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optional_vars.R
\docType{data}
\name{optional_vars}
\alias{optional_vars}
\title{Optional Variables}
\format{
Character vector of variables that are optional. Includes most scores.
}
\usage{
optional_vars
}
\description{
Optional Variables
}
\keyword{datasets}
