% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotModules.R
\name{plotApp}
\alias{plotApp}
\title{Plot app}
\usage{
plotApp(dat_input = prepare_data(demo_data), studyids = NULL)
}
\arguments{
\item{dat_input}{A data frame. Defaults to \code{prepare_data(demo_data)}.}

\item{studyids}{Optional. If \code{NULL} (default), will use unique values from \code{dat_input$NACCID}.}
}
\value{
A shiny app.
}
\description{
Shiny app using the plotUI and plotServer.
}
