% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_new_traces.R
\name{plotly_new_traces}
\alias{plotly_new_traces}
\title{New Traces to add to Plot}
\usage{
plotly_new_traces(
  new_dat,
  visibility = visibility_defaults,
  legend_names,
  vars_colors
)
}
\arguments{
\item{new_dat}{A data frame.}

\item{visibility}{A named list, where the names correspond to variables in
\code{new_dat} and the values are one of \code{"TRUE"}, \code{"FALSE"}, or \code{"legendonly"}.}

\item{legend_names}{Legend names. Will be returned with added legends if any
new ones were created. This only happens if a pair of crosswalk variables
are both present.}

\item{vars_colors}{Named vector. Entries give colors to use for markers and liens
for variable corresponding to the name of the entry.}
}
\value{
A list containing new traces (a list of lines and markers), legend names, the new
x range, and the new y range. If there are no new traces to add (i.e. no standardized
scores in the data), the return is empty.
}
\description{
A short description...
}
