% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotModules.R
\name{plotUI}
\alias{plotUI}
\title{Plot UI}
\usage{
plotUI(id)
}
\arguments{
\item{id}{A single string to ID the Shiny module, which also indicates
which group of variables to plot. Should be one of \code{unique(nacc_var_groups)}.}
}
\value{
A \code{shiny::tagList} object containing two \code{shiny::conditionalPanel}'s:
one with the plot, and one with text for when no standardized scores
were found. Both wrapped in a \code{bslib::accordion_panel} for inclusion in
main app.
}
\description{
UI for displaying plots of standardized scores across visits.
}
