% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVarModule.R
\name{plotVarUI}
\alias{plotVarUI}
\alias{plotVarServer}
\alias{plotVarApp}
\title{Shiny Module to Plot Scores Across Visits}
\usage{
plotVarUI(id)

plotVarServer(
  id,
  dat,
  studyid,
  descriptions = c(Impaired = 0.03, Borderline = 0.1, `Low Average` = 0.26, Average =
    0.76, `High Average` = 0.92, Superior = 0.97, `Very Superior` = 1),
  fill_values = NULL,
  print_updating = T
)

plotVarApp(dat = prepare_data(demo_data), studyids = NULL)
}
\arguments{
\item{id}{An ID string to match module UI and module server}

\item{dat}{Data to use for the plots.}

\item{studyid}{ID of participant; reactive}

\item{descriptions}{A named vector giving cutoffs to use for creating descriptions.}

\item{fill_values}{(optional) A named vector of same length (with same names)
as \code{descriptions} with hex color values to use. If \code{NULL}, no colorcoding used.
By default, evenly spread out colors from red through yellow to green are used.}

\item{print_updating}{logical (defualt \code{FALSE}); should a message be displayed when table is being updated? For debugging.}

\item{studyids}{vector of studyids to include in drop-down menu}
}
\description{
This shiny module is meant for visualizing changes over time of a long list
of cognitive measures from the WLS ILIAD study.
}
\examples{
\dontrun{
plotVarApp(demo_data[1, ])
}

}
