% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Wrapper to prepare raw data}
\usage{
prepare_data(
  dat,
  selected_cols = c("RACE", "CDRGLOB", "MOCATOTS", "MOCBTOTS", "TRAILA", "TRAILARR",
    "TRAILALI", "OTRAILA", "OTRLARR", "DIGFORCT", "DIGFORSL", "DIGBACCT", "DIGBACLS",
    "WAIS", "MINTTOTS", "ANIMALS", "VEG", "UDSVERTN", "UDSVERFC", "UDSVERLC", "UDSBENTC",
    "UDSBENTD", "CRAFTVRS", "CRAFTURS", "CRAFTDVR", "CRAFTDRE", "REY1REC", "REY2REC",
    "REY3REC", "REY4REC", "REY5REC", "REY6REC", "REYDREC", "REYTCOR", "TRAILB",
    "TRAILBLI", "TRAILBRR", "MOCACLOC", "MOCACLOH", "MOCACLON", "OTRAILB", "OTRLBRR",
    "OTRLBLI", "NACCGDS", "CDRSUM", 
     "UDSBENRS", "NACCID", "SEX", "EDUC", "NACCAGE",
    "BIRTHYR", "VISITYR", "VISITMO", "VISITDAY", "NACCUDSD", "NACCMMSE", "BOSTON",
    "LOGIMEM", "MEMUNITS", "MEMTIME", "DIGIF", "DIGIFLEN", "DIGIB", "DIGIBLEN"),
  methods = NULL,
  print_messages = F,
  with_diags = T
)
}
\arguments{
\item{dat}{data set similar to the NACC data. For an example, see
\code{?demo_data}.}

\item{selected_cols}{vector with columns to keep. If named, the names must
correspond to NACC variable names and entries columns in the data set \code{dat}.}

\item{methods}{NULL (default) or list of named entries specifying which model to use
for standardizing cognitive scores. If NULL, defaults are applied for all
variables in the dataset for which methods have been implemented (see \code{NpsychBatteryNorms::default_methods})}

\item{print_messages}{logical; should messages be printed? If TRUE (default), a message will
be print with the methods used if methods is omitted (or NULL)}

\item{with_diags}{logical (default \code{TRUE}); should diagnoses be included?}
}
\description{
This is a wrapper function that returns only the desired columns, adds
variables that can be derived to the data, and adds standardized scores.
}
\examples{
prepare_data(demo_data)

}
