% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prev_diagnoses_table.R
\name{prev_diagnoses_table}
\alias{prev_diagnoses_table}
\title{Table With Previous Diagnoses}
\usage{
prev_diagnoses_table(dat, table_font_size = 100)
}
\arguments{
\item{dat}{Data to use. Should only refer to a single participant}

\item{table_font_size}{Table font size as a percent. Default: 100}
}
\description{
Table With Previous Diagnoses
}
