% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevDiagnosesModule.R
\name{prevDiagnosesUI}
\alias{prevDiagnosesUI}
\alias{prevDiagnosesServer}
\alias{prevDiagnosesApp}
\title{Shiny Module to Display Previous Diagnoses}
\usage{
prevDiagnosesUI(id)

prevDiagnosesServer(
  id,
  dat,
  print_updating = F,
  table_font_size = shiny::reactive(100)
)

prevDiagnosesApp(dat)
}
\arguments{
\item{id}{An ID string to match module UI and module server}

\item{dat}{Data to use. Should only refer to a single participant}

\item{print_updating}{logical (default \code{FALSE}); should a message be printed to indicate that the table is being updated? For debugging.}

\item{table_font_size}{Table font size as a percent. Default: 100}
}
\description{
Shiny Module to Display Previous Diagnoses
}
