% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdd.R
\docType{data}
\name{rdd}
\alias{rdd}
\title{Researchers Data Dictionary in List Form}
\format{
An object of class \code{list} of length 776.
}
\usage{
rdd
}
\description{
Researchers Data Dictionary in List Form
}
\keyword{datasets}
