% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{Read data}
\usage{
read_data(data_source = "redcap", data_type = NULL, redcap_auth, data_file)
}
\arguments{
\item{data_source}{One of \code{"redcap"}, \code{"csv_upload"}, or \code{"demo"}.}

\item{data_type}{Optional. If \code{data_source} is \code{"redcap"}, then must be one of \code{"wadrc_uds2"}, \code{"wadrc_uds3"} or \code{"wadrc_uds4"}.}

\item{redcap_auth}{Optional. A list with entries \code{redcap_uri} and \code{token} (when \code{data_source = "redcap"}).}

\item{data_file}{Optional. A character vector with path to data file (when \code{data_source = "csv"}).}
}
\value{
A \code{data.table} object.
}
\description{
Function used to read in data based on the output returned from \code{dataSelectServer()}.
}
