% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_labels.R
\name{var_labels}
\alias{var_labels}
\title{Create Labels for Assessment Summary Table}
\usage{
var_labels(
  CDRSUM = NA,
  TRAILARR = NA,
  TRAILALI = NA,
  TRAILBRR = NA,
  TRAILBLI = NA,
  UDSBENRS = NA,
  UDSBENTD = NA,
  UDSBENTC = NA,
  REY1REC = NA,
  REY2REC = NA,
  REY3REC = NA,
  REY4REC = NA,
  REY5REC = NA,
  REYTCOR = NA,
  REYFPOS = NA,
  CRAFTURS = NA,
  CRAFTVRS = NA,
  CRAFTDVR = NA,
  CRAFTDRE = NA
)
}
\arguments{
\item{CDRSUM}{Standard CDR sum of boxes}

\item{TRAILARR}{Number of commission errors for Trail Making Part A}

\item{TRAILALI}{Number of correct lines for Trail Making Part A}

\item{TRAILBRR}{Number of commission errors for Trail Making Part B}

\item{TRAILBLI}{Number of correct lines for Trail Making Part B}

\item{UDSBENRS}{Benson Complex Figure Recall - Recognized original stimulus among four options}

\item{UDSBENTD}{Total score for 10- to 15-minute delayed drawing of Benson figure}

\item{UDSBENTC}{Total Score for copy of Benson figure}

\item{REY1REC}{Rey AVLT Trail 1 total recall}

\item{REY2REC}{Rey AVLT Trail 2 total recall}

\item{REY3REC}{Rey AVLT Trail 3 total recall}

\item{REY4REC}{Rey AVLT Trail 4 total recall}

\item{REY5REC}{Rey AVLT Trail 5 total recall}

\item{REYTCOR}{Rey AVLT delayed recognition; total correct}

\item{REYFPOS}{Rey AVLT delayed recognition; total false positives}

\item{CRAFTURS}{Craft Story 21 Recall (Immediate) - Total story units recalled, paraphrase scoring}

\item{CRAFTVRS}{Craft Story 21 Recall (Immediate) - Total story units recalled, verbatim scoring}

\item{CRAFTDVR}{Craft Story 21 Recall (Delayed) - Total story units recalled, verbatim scoring}

\item{CRAFTDRE}{Craft Story 21 Recall (Delayed) - Total story units recalled, paraphrase scoring}
}
\description{
Helper to create row names for \code{assessment_summary_table} based on values of
various scores.
}
\examples{
var_labels(
  CDRSUM = 14,
  TRAILARR = 0,
  TRAILALI = 24,
  TRAILBRR = 2,
  TRAILBLI = 24,
  UDSBENRS = 1,
  UDSBENTD = 14,
  UDSBENTC = 14,
  REY1REC = -4,
  REY2REC = -4,
  REY3REC = -4,
  REY4REC = -4,
  REY5REC = -4,
  REYTCOR = -4,
  REYFPOS = -4,
  CRAFTURS = 20,
  CRAFTVRS = 12,
  CRAFTDVR = 11,
  CRAFTDRE = 13
)

}
