% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visibility_defaults.R
\docType{data}
\name{visibility_defaults}
\alias{visibility_defaults}
\title{Visibility Defaults}
\format{
List with named entries giving the default visibility setting for
the variables to use in the plots.
}
\usage{
visibility_defaults
}
\description{
Visibility Defaults
}
\keyword{datasets}
