% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visit_vars.R
\docType{data}
\name{visit_vars}
\alias{visit_vars}
\title{Variables necessary for Visit Dates}
\format{
Character vector of variables used for determining visit dates.
}
\usage{
visit_vars
}
\description{
Variables necessary for Visit Dates
}
\keyword{datasets}
