% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wadrc_data_prep.R
\name{wadrc_data_prep}
\alias{wadrc_data_prep}
\title{Function to Prepare WADRC Data}
\usage{
wadrc_data_prep(adrc_data)
}
\arguments{
\item{adrc_data}{Data as gotten from WADRC REDCap.
Could be a complete download to .csv file, then read into R,
or pulled directly from REDCap using the \code{REDCapR} package.}
}
\description{
Function to Prepare WADRC Data
}
