% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wadrc_redcap_fields.R
\docType{data}
\name{wadrc_redcap_fields}
\alias{wadrc_redcap_fields}
\title{Fields to pull from WADRC REDCap}
\format{
Character vector with names of fields we pull from the WADRC REDCap database
}
\usage{
wadrc_redcap_fields
}
\description{
Fields to pull from WADRC REDCap
}
\keyword{datasets}
