% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wadrc_to_nacc.R
\docType{data}
\name{wadrc_to_nacc}
\alias{wadrc_to_nacc}
\title{Translate WADRC variable names to NACC variable names}
\format{
A named vector where names of entries correspond to WADRC variable
names, and entries contain NACC variable names.
}
\usage{
wadrc_to_nacc
}
\description{
Translate WADRC variable names to NACC variable names
}
\keyword{datasets}
