% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wls_to_nacc.R
\docType{data}
\name{wls_to_nacc}
\alias{wls_to_nacc}
\title{Translate WLS variable names to NACC variable names}
\format{
A named vector where names of entries correspond to WLS variable
names, and entries contain NACC variable names.
}
\usage{
wls_to_nacc
}
\description{
Translate WLS variable names to NACC variable names
}
\keyword{datasets}
