% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarkerModule.R
\name{biomarkerUI}
\alias{biomarkerUI}
\alias{biomarkerServer}
\alias{biomarkerApp}
\title{Biomarker UI}
\usage{
biomarkerUI(id)

biomarkerServer(id, adrc_ptid, biomarker_api)

biomarkerApp(adrc_ptid, biomarker_api, testing = FALSE)
}
\arguments{
\item{id}{A string used to namespace the module.}

\item{adrc_ptid}{ADRC ptid. A character vector with patient ID's that can be chosen.}

\item{biomarker_api}{Panda API key.}
}
\value{
A UI definition.

NULL.

A \code{shinyApp} object.
}
\description{
A short description...

A short description...

A short description...
}
\keyword{internal}
