% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colSelectModule.R
\name{colSelectUI}
\alias{colSelectUI}
\alias{colSelectServer}
\alias{colSelectApp}
\title{Shiny Module for Selecting Columns}
\usage{
colSelectUI(id)

colSelectServer(
  id,
  col_names,
  default_methods,
  data_type,
  allow_col_selection = T
)

colSelectApp(
  col_names,
  default_methods,
  data_type = c("nacc", "wls", "wadrc"),
  allow_col_selection = T
)
}
\arguments{
\item{id}{id to link shiny modules}

\item{col_names}{reactive character vector with names to choose from}

\item{default_methods}{named list. Each entry should be named after a variable. The entry should be a named character vector with to elements: one named 'method' to indicate the standardization method to use, and one name 'version' to indicate the version to use for standardization.}

\item{data_type}{reactive value; one of 'wls', 'nacc', 'wadrc'}

\item{allow_col_selection}{logical; if \code{TRUE} (default), allow user to select which columns should be used for each variable.}
}
\description{
This shiny module gives the user the option to manually map column names to
NACC variable names.
}
