% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSelectModule.R
\name{dataSelectUI}
\alias{dataSelectUI}
\alias{dataSelectServer}
\alias{dataSelectApp}
\title{Shiny module to select data for assessment table}
\usage{
dataSelectUI(id)

dataSelectServer(id)

dataSelectApp(testing = FALSE)
}
\arguments{
\item{id}{A string used to create a namespace within the shiny app.}

\item{testing}{Logical, whether to run the app in testing mode.}
}
\value{
A UI definition.

A list containing reactive values for the data object (\code{dat_obj}),
the data source, the data type, and the biomarker API.
}
\description{
A short description...

Server logic for dataSelectModule.
}
