% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptionsModule.R
\name{descriptionsUI}
\alias{descriptionsUI}
\alias{descriptionsServer}
\alias{descriptionsApp}
\title{Descriptions UI}
\usage{
descriptionsUI(id)

descriptionsServer(
  id,
  default_descriptions = c(Impaired = 0.03, Borderline = 0.1, `Low Average` = 0.26,
    Average = 0.76, `High Average` = 0.92, Superior = 0.97, `Very Superior` = 1),
  default_fill_values = NULL
)

descriptionsApp(
  default_descriptions = c(Impaired = 0.03, Borderline = 0.1, `Low Average` = 0.26,
    Average = 0.76, `High Average` = 0.92, Superior = 0.97, `Very Superior` = 1),
  testing = FALSE
)
}
\arguments{
\item{id}{A string.}

\item{default_descriptions}{Optional. A named numeric vector with names for \code{Label} and values for \code{Upper_Bound}.}

\item{default_fill_values}{Optional. A named character vector with names for \code{Label} and values for colors (as HEX values)}

\item{testing}{Logical, whether to run the app in testing mode.}
}
\value{
A UI definition.

A list containing \code{fill_values} and \code{descriptions}.
}
\description{
UI to let the user select the groups used for coloring. Includes a table where
group labels can be specified, cutoffs set, and color selected.

A short description...
}
