% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NpsychAssessmentTool-package.R
\docType{package}
\name{NpsychAssessmentTool-package}
\alias{NpsychAssessmentTool}
\alias{NpsychAssessmentTool-package}
\title{NpsychAssessmentTool: UDS Battery Scores in Interactive Tables and Figures}
\description{
Building on the NpsychBatteryNorms package, this package includes a Shiny application that makes it easy to present UDS battery scores in a way useful to consensus diagnosis. Data can be imported as an .csv file or pulled directly from a REDCap database (`REDCapR` needed).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rmtrane/NpsychAssessmentTool}
  \item Report bugs at \url{https://github.com/rmtrane/NpsychAssessmentTool/issues}
}

}
\author{
\strong{Maintainer}: Ralph Møller Trane \email{rtrane@wisc.edu}

}
\keyword{internal}
